requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bleach==4.1.0
Django==4.0.2
django-braces==1.15.0
django-crispy-forms==1.14.0
django-etc==1.3.1
django-filter==21.1
django-location-field==2.1.0
django-model-utils==4.2.0
django-summernote==0.8.20.0
django-utils-six==2.0
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-gettext==4.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tzdata==2021.5
webencodings==0.5.1
